feat: Add Firebase dependency and authentication context
This commit adds the Firebase dependency to the project by including it in the package.json file. Additionally, it introduces an authentication context by creating the AuthContext component and using it in the App component. This will allow for user authentication and authorization functionality to be implemented in the future.